Add files via upload
The .docx file is a summary of findings, and is filled with data visualizations, statistical tables, and a brief summary of the support vector machine (SVM) model's results in predicting loan defaults. The PD_Calculator.py can be used to estimate probabilities of default for peer-to-peer loans with 80% accuracy. The pickle file is the 80% accurate SVM model that was trained on 1300 loans. The PD_Calculator needs the pickle file to work, as it loads it. PD_project.py was the file used to make the SVM model. The PD_project.py was also used to generate the statistical tables and data visualizations that are in the .docx report.

NOTE: I could not upload the loan data onto GitHub, as it easily exceeds the file size limit. You can download the dateset of over 1 million loans here: https://www.kaggle.com/wendykan/lending-club-loan-data?select=loan.csv